% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clocus_lower.R
\name{clocus_lower}
\alias{clocus_lower}
\title{A clocus_lower Funtion.}
\usage{
clocus_lower(gene_GWAS, gene_chr, gene_everyline = 6, x_title,
  grid.show = element_blank(), omit_gene = c())
}
\arguments{
\item{gene_GWAS}{The Gene GWAS file.}

\item{gene_chr}{Chromosome specific gene list file}

\item{gene_everyline}{The maximum number of genes in every line, the default number is 6.}

\item{grid.show}{Show grid or not, default is False.}

\item{omit_gene}{The genes you do not want to show in the figure. Default is a empty list.}

\item{top_snp}{Name of top snp}
}
\description{
This function allows you to plot th lower plot in the example.
}
\examples{
all_gene<-fread("gencode.v19.withdir.txt",header=T)
C1QTNF4<-fread("C1QTNF4.txt",header=T)
#LD file
C1QTNF4_ld<-fread("C1QTNF4.ld",header=T)
C1QTNF4<-C1QTNF4[order(C1QTNF4$POS)]
C1QTNF4_ld<-C1QTNF4_ld[order(C1QTNF4_ld$BP_B)]
C1QTNF4<-C1QTNF4[C1QTNF4$SNP \%in\% C1QTNF4_ld$SNP_B]
#combine the LD dataset with GWAS dataset =>gene_GWAS
C1QTNF4_p<-data.frame(P=C1QTNF4$P,r2=C1QTNF4_ld$R2,snp=C1QTNF4$SNP,pos=C1QTNF4$POS)
top_p =9.702e-11
top_pos=47380340
top_snp="rs3740688"
all_gene_11<-all_gene[all_gene$chr==11]
g2<-clocus_lower(C1QTNF4_p,all_gene_11,x_title = "chromosome 11 position (Mb)")
g2
}
\keyword{lower}
