% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clocus_upper.R
\name{clocus_upper}
\alias{clocus_upper}
\title{A clocus_upper Funtion.}
\usage{
clocus_upper(gene_GWAS, grid.show = element_line(color = "grey94"),
  top_pos, top_p, top_snp, color = c("#c1d8e6", "#537ea3", "#d0b8d9",
  "#a283a3", "#739e6f", "#b3d49b", "#ffc887", "#e3987f", "#ffb5b5",
  "#bf4d4d"))
}
\arguments{
\item{gene_GWAS}{The Gene GWAS file.}

\item{grid.show}{Show grid or not, default is False.}

\item{top_pos}{Position of top SNP you want to highlight}

\item{top_p}{P value of top snp}

\item{top_snp}{Name of top snp}

\item{color}{Color used to plot the points. 10 levels.}
}
\description{
This function allows you to plot th upper plot in the example.
}
\examples{
all_gene<-fread("gencode.v19.withdir.txt",header=T)
C1QTNF4<-fread("C1QTNF4.txt",header=T)
#LD file
C1QTNF4_ld<-fread("C1QTNF4.ld",header=T)
C1QTNF4<-C1QTNF4[order(C1QTNF4$POS)]
C1QTNF4_ld<-C1QTNF4_ld[order(C1QTNF4_ld$BP_B)]
C1QTNF4<-C1QTNF4[C1QTNF4$SNP \%in\% C1QTNF4_ld$SNP_B]
#combine the LD dataset with GWAS dataset =>gene_GWAS
C1QTNF4_p<-data.frame(P=C1QTNF4$P,r2=C1QTNF4_ld$R2,snp=C1QTNF4$SNP,pos=C1QTNF4$POS)
top_p =9.702e-11
top_pos=47380340
top_snp="rs3740688"
all_gene_11<-all_gene[all_gene$chr==11]
g1<-clocus_upper(C1QTNF4_p,grid.show = element_blank(),top_p = top_p,top_pos = top_pos,top_snp = top_snp)
g1
}
\keyword{upper}
